feat(table): rename GenericTable component to BIMDataTable (replacing old table)
