+++ src/BIMDataComponents/BIMDataTable/_BIMDataTable.scss
@@ -1,40 +1,57 @@
 .bimdata-table {
   width: 100%;
-  font-size: calculateEm(14px);
-  table {
-    width: 100%;
-    border-collapse: collapse;
-    font-family: $primary-font;
-    border-spacing: 0;
-    tr {
+  font-size: 0.9rem;
+  background-color: $color-white;
+
+  &__container {
+    table {
+      width: 100%;
+      border-collapse: collapse;
+      border-spacing: 0;
+
+      tbody tr:nth-child(odd) {
+        background-color: $color-tertiary-lightest;
+      }
+
       th,
       td {
-        height: 50px;
+        min-height: $spacing-unit * 2;
         padding: 0 $spacing-unit;
-        min-height: calc(#{$spacing-unit} * 2);
-        font-size: calculateEm(11px);
-        text-align: left;
+
+        &.cell-checkbox:deep() {
+          width: 42px;
+        }
+      }
+
+      th {
+        font-weight: bold;
+        color: $color-primary;
+      }
+
+      td {
+        color: $color-tertiary-darkest;
       }
     }
-    tbody {
-      tr {
-        td {
-          color: $color-tertiary-darkest;
-        }
-        &:nth-child(odd) {
-          background-color: $color-tertiary-lightest;
-        }
-      }
+
+    &__placeholder {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      padding: $spacing-unit;
+      font-size: 1rem;
+      font-weight: bold;
+      background-color: $color-tertiary-lightest;
     }
   }
-  &__compensated {
-    table {
-      tr {
-        th,
-        td {
-          height: 35px;
-        }
-      }
+
+  &__page-nav {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    padding: $spacing-unit;
+
+    &__text {
+      margin: 0 $spacing-unit;
     }
   }
 }
